Create PUT methods, remove old README
The old README was still around, we didn't need that.

This adds PUT methods so that we can hopefully delete our tasks
at some point. We really need the CSRF token in order for PUT
requests to work, though. Until we have that, they error out
with a message that leads me to believe that the API itself
acts weird and returns an invalid HTTP response. I can reproduce
it with cURL if I make the request without the CSRF header.