Add CSRF cookie/header
When we make PUT requests, we have to include the CSRF token.
The Echo CSRF system apparently works by setting a cookie called
`csrf`, and also setting a header called `csrf`. If the two match,
the CSRF is considered to be valid.

That said, PUT requests still don't work. They are still responded
to with a Max Retries Exceeded, BadStatusLine exception. Haven't
figured that part out yet.